pyproject.toml
[project]
name = "next-door-key-simulator"
version = "2.0.0"
description = "Quantum key distribution (QKD) key management system (KMS) simulator based on the ETSI GS QKD 014 V1.1.1 standard"
requires-python = '>=3.10'

dependencies = [
    "blinker>=1.6.2",
    "certifi>=2023.7.22",
    "cffi>=1.15.1",
    "charset-normalizer>=3.2.0",
    "click>=8.1.7",
    "cryptography>=41.0.4",
    "Flask>=2.3.3",
    "idna>=3.4",
    "itsdangerous>=2.1.2",
    "Jinja2>=3.1.2",
    "MarkupSafe>=2.1.3",
    "packaging>=23.1",
    "pycparser>=2.21",
    "pyOpenSSL>=23.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "urllib3>=2.0.5",
    "Werkzeug>=2.3.7"
]